Fix flaky test when reconciler start before previous one still up
between test, the reconciler is stopped and started again.
Stop is not instant but we started a new instance right away.
Since it bind a port, the new one crashed because the old
one didn't release it yet. Instead we wait for the manager to
be properly down before starting a new one