Improve deadline exceed message (#65)
* Improve deadline exceed message

Deadline exceeded is what the status of a node disruption will
show once a Node Disruption is fully rejected. Most of the time,
we reach the deadline because some budgets are preventing
from reaching the granted state. In the previous version, we lost
the context and had to look at logs. Now we keep the previous
status and only append a reason. This way we know why the last
attempt was rejected.

* Refactor internal disruption status to avoid shallow functions

* Refactor to remove duplication of code

Make the reference calculation a method from ndr
Remove redondant error handling for RejectOverlappingDisruption